fix : Добавление проверок в ServiceDirector.java

+++ src/main/java/ru/yandex/practicum/filmorete/service/ServiceDirector.java
@@ -9,6 +9,7 @@ import ru.yandex.practicum.filmorete.exeptions.ExceptionNotFoundDirectorStorage;
 import ru.yandex.practicum.filmorete.model.Director;
 import ru.yandex.practicum.filmorete.sql.dao.DirectorDao;
 
+import static ru.yandex.practicum.filmorete.exeptions.message.DirectorErrorMessage.SERVICE_ERROR_DIRECTOR_DUPLICATE_IN_COLLECTIONS;
 import static ru.yandex.practicum.filmorete.exeptions.message.DirectorErrorMessage.SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS;
 
 @Service
@@ -25,8 +26,8 @@ public class ServiceDirector {
      */
     public Director getDirectorSearchId(Long directorId) {
         Optional<Director> result = directorDao.findById(directorId);
-        if (result.isPresent()) return result.get();
-        else throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS);
+        if (result.isPresent()) throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS);
+        return result.get();
     }
 
     /**
@@ -40,6 +41,8 @@ public class ServiceDirector {
      * Добавление нового режиссёра [ DIRECTORS ].
      */
     public Director add(@NotNull Director director) {
+        Optional<Director> optionalDirector = directorDao.findByName(director.getName());
+        if (optionalDirector.isPresent()) throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_DUPLICATE_IN_COLLECTIONS);
         Long id = directorDao.insertByName(director.getName());
         Optional<Director> result = directorDao.findById(id);
         return result.orElse(null);
@@ -50,11 +53,10 @@ public class ServiceDirector {
      */
     public Director update(@NotNull Director director) {
         Optional<Director> optionalDirector = directorDao.findById(director.getId());
-        if (optionalDirector.isPresent()) {
-            directorDao.update(director.getId(), director.getName());
-            Optional<Director> result = directorDao.findById(director.getId());
-            return result.orElse(null);
-        } else throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS);
+        if (optionalDirector.isEmpty()) throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS);
+        directorDao.update(director.getId(), director.getName());
+        Optional<Director> result = directorDao.findById(director.getId());
+        return result.orElse(null);
     }
 
     /**
@@ -68,6 +70,8 @@ public class ServiceDirector {
      * Удаление режиссёра по ID [ DIRECTORS ].
      */
     public void deleteSearchId(Long directorId) {
+        Optional<Director> optional = directorDao.findById(directorId);
+        if (optional.isEmpty()) throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS);
         directorDao.deleteById(directorId);
     }
 
@@ -75,6 +79,8 @@ public class ServiceDirector {
      * Удаление режиссёра по имени [ DIRECTORS ].
      */
     public void deleteSearchName(String directorName) {
+        Optional<Director> optional = directorDao.findByName(directorName);
+        if (optional.isEmpty()) throw new ExceptionNotFoundDirectorStorage(SERVICE_ERROR_DIRECTOR_NOT_IN_COLLECTIONS);
         directorDao.deleteByName(directorName);
     }
 }
